EPMRPP-79630 || Code review fixes - 1
