+++ app/src/components/widgets/singleLevelWidgets/charts/common/passingRateChart/config/utils.js
@@ -18,9 +18,6 @@ import { CHART_MODES, MODES_VALUES } from 'common/constants/chartModes';
 
 export const getPercentage = (value, totalItems) => ((value / totalItems) * 100).toFixed(2);
 
-export const getPercentageWithoutSkipped = (value, totalItemsWithoutSkipped) =>
-  getPercentage(value, totalItemsWithoutSkipped);
-
 export const getChartViewModeOptions = (viewMode, isPreview, totalItems) =>
   viewMode === MODES_VALUES[CHART_MODES.BAR_VIEW]
     ? {
@@ -66,12 +63,6 @@ export const getChartViewModeOptions = (viewMode, isPreview, totalItems) =>
         },
       };
 
-export const getChartViewModeOptionsWithoutSkipped = (
-  viewMode,
-  isPreview,
-  totalItemsWithoutSkipped,
-) => getChartViewModeOptions(viewMode, isPreview, totalItemsWithoutSkipped);
-
 export const calculateTooltipParams = (data, color, customProps) => {
   const { totalItems, formatMessage, includeSkipped, totalItemsWithoutSkipped } = customProps;
   const { id, name, value } = data[0];
@@ -80,7 +71,7 @@ export const calculateTooltipParams = (data, color, customProps) => {
     itemsCount: `${value} (${
       includeSkipped
         ? getPercentage(value, totalItems)
-        : getPercentageWithoutSkipped(value, totalItemsWithoutSkipped)
+        : getPercentage(value, totalItemsWithoutSkipped)
     }%)`,
     color: color(id),
     issueStatNameProps: { itemName: name, defectTypes: {}, formatMessage },
