+++ app/src/pages/inside/dashboardItemPage/modals/common/widgetControls/controls/radioGroupControl.scss
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2019 EPAM Systems
+ * Copyright 2022 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-.radioGroupControl-wrapper {
+.radio-group-control-wrapper {
   padding-top: 10px;
-  .radioGroup:not(:last-child) {
+  .radio-group:not(:last-child) {
     padding-bottom: 11px;
   }
 }
